+++ groovyConsole/history.md
@@ -1,5 +1,12 @@
 # History
 
+## v0.0.2
+
+* Added "GroovyScript plugin" 's compiler configuration:
+It adds access to Freeplane's and add-ons libraries
+* Added button to refresh Compiler Configuration
+* Added console.systemOutInterceptor.start() to capture print() and println() in the output console or the log.0 file
+
 ## v0.0.1
 
 * Initial release